Squashed commit of the following:
commit c487f686cb59cd6d025a7d2f4dc0c719f72920c2
Merge: f671a04 71b14ca
Author: Paula <paulabomp@gmail.com>
Date:   Thu Jun 5 17:45:25 2025 -0300

    Merge branch 'Pau🤠'

commit f671a04869f4c03d4cf006b1ec65a074b193465b
Merge: 98200ac ebf736e
Author: Ramiro <deconinckramiro@gmail.com>
Date:   Wed Jun 4 18:19:25 2025 -0300

    Merge branch 'ramiro'

commit ebf736e6b2b7ca1cc102303e555f1273ee1f019e
Author: Ramiro <deconinckramiro@gmail.com>
Date:   Wed Jun 4 18:17:57 2025 -0300

    Testeo y corrección CU ListarArtXProv

commit 2945d91033294a18e1a9322fd44ab5b7990f2d64
Author: Ramiro <deconinckramiro@gmail.com>
Date:   Wed Jun 4 18:04:44 2025 -0300

    CU ListarArticuloXProveedor

commit 98200ac67bb54d0bfe269da412b09270dda4547b
Merge: ab7af0f 54a3ef0
Author: Ramiro <deconinckramiro@gmail.com>
Date:   Tue Jun 3 23:27:53 2025 -0300

    Merge remote-tracking branch 'origin/ramiro'

commit ab7af0f38b1a056ca7acd7bab74931faaf44b69c
Author: Ramiro <deconinckramiro@gmail.com>
Date:   Tue Jun 3 23:24:59 2025 -0300

    Correccion archivo main

commit 54a3ef02e2c4bd19d0487a44fa5769ef45da53e1
Author: Ramiro <deconinckramiro@gmail.com>
Date:   Tue Jun 3 23:21:28 2025 -0300

    Correcciones version anterior

commit 69c17f0ee42f242f446fae07063338af4d55b0af
Merge: 67fa7f7 d803b35
Author: Ramiro <deconinckramiro@gmail.com>
Date:   Tue Jun 3 22:01:46 2025 -0300

    corrección error en el merge

commit 67fa7f7c40db86bd147936ba6585fa33b5edc280
Author: Ramiro <deconinckramiro@gmail.com>
Date:   Tue Jun 3 21:42:02 2025 -0300

    Cambios necesarios en entidades OC y OCDetalle

commit 9fb6f3debc3542db43f809a2ca33be7fd8c2c6d2
Author: Ramiro <deconinckramiro@gmail.com>
Date:   Tue Jun 3 21:41:19 2025 -0300

    Arreglos CUProponerOC y testeo

commit d803b35f6c696ed75b013279db0c4c5eb8e388d3
Author: Renzo Sahade <ing.sahade.renzo@gmail.com>
Date:   Tue Jun 3 11:44:49 2025 -0300

    Finalizado CU Generar Venta

commit d408333cccfd926547e4163c69fa44a6eae5a789
Author: Renzo Sahade <ing.sahade.renzo@gmail.com>
Date:   Tue Jun 3 10:59:44 2025 -0300

    CU Generar Venta

commit 5acacc63cc3d911817ba359e9fe1c4e7d8e52dd6
Author: Renzo Sahade <ing.sahade.renzo@gmail.com>
Date:   Mon Jun 2 23:16:08 2025 -0300

    CU Asignar proveedor

commit 59fa3ec356850a47e7fbc672184fb8c4f7b999ee
Author: Renzo Sahade <ing.sahade.renzo@gmail.com>
Date:   Mon Jun 2 23:15:48 2025 -0300

    Modificar ArticuloProveedor - Finalizando CU Asignar proveedor

commit 373358514e74cd237ba1dc62700ab0b3d922f421
Author: Renzo Sahade <ing.sahade.renzo@gmail.com>
Date:   Mon Jun 2 23:08:59 2025 -0300

    Eliminar ArticuloProveedor

commit 905ccff0c0360b6280badeaf59fdb56c611213fe
Author: Renzo Sahade <ing.sahade.renzo@gmail.com>
Date:   Mon Jun 2 22:57:46 2025 -0300

    Alta ArticuloProveedor

commit 14219031507220351cb702eac6aa19be707dbaf8
Author: Renzo Sahade <ing.sahade.renzo@gmail.com>
Date:   Mon Jun 2 20:07:17 2025 -0300

    Finalizando ABM Articulo

commit 3b3edbcc11b4ad02a5b26b49447b2be56ad22d2a
Merge: 4b8d134 59cfe94
Author: Renzo Sahade <ing.sahade.renzo@gmail.com>
Date:   Mon Jun 2 19:37:22 2025 -0300

    Merge

commit 59cfe946cdcc7a60b8cd7327fdd6018f183959c4
Merge: 57a0e03 39182a0
Author: Renzo Sahade <ing.sahade.renzo@gmail.com>
Date:   Mon Jun 2 19:35:22 2025 -0300

    Merge branch 'main' of https://github.com/Facustriker/InvOperativa-BackEnd

commit 4b8d1346fa87bdc171d39c29ae0561ed931230d4
Author: Renzo Sahade <ing.sahade.renzo@gmail.com>
Date:   Mon Jun 2 19:35:10 2025 -0300

    Commit auxiliar

commit 8f18611c97349d63486fbf1569fd75b4e33b4b93
Author: Ramiro <deconinckramiro@gmail.com>
Date:   Sun Jun 1 23:32:53 2025 -0300

    CU19 ProponerOrdenCompra

commit 57a0e0372d135aabc68f33e70e9875afd7e3de47
Author: Renzo Sahade <ing.sahade.renzo@gmail.com>
Date:   Mon May 26 00:00:51 2025 -0300

    Corrigiendo ABM Articulo

+++ src/main/java/com/invOperativa/Integrador/Entidades/DetalleVenta.java
@@ -16,7 +16,7 @@ public class DetalleVenta extends BaseEntity{
     private int cant;
 
     @Column(name = "subTotal")
-    private int subTotal;
+    private float subTotal;
 
     @ManyToOne
     @JoinColumn(name = "articulo")
